backend/requirements.txt
absl-py==2.1.0
albumentations==1.4.20
anyio==4.6.2.post1
attrs==24.2.0
bokeh==3.6.1
fastapi==0.115.5
fastapi-cli==0.0.5
fire==0.7.0
h11==0.14.0
h5py==3.12.1
jinja2==3.1.4
joblib==1.4.2
json-tricks==3.17.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mlflow==2.17.2
monai==1.4.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
optuna==4.1.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
plotly==5.24.1
protobuf==5.28.3
psutil==6.1.0
pydantic==2.9.2
pydicom==3.0.1
pygments==2.18.0
pynrrd==1.1.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.3
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
sqlalchemy==2.0.36
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
uvicorn==0.32.1
Werkzeug==3.1.3
nibabel
torchxrayvision
umap
pycaret
catboost
xgboost
seaborn
scanpy
hdf5plugin
scanorama
decoupler